compute sum_uncovered_weight in greedy_construct

+++ npbenchmark-main/SDK.PCP/pcp_vector.h
@@ -58,7 +58,7 @@ private: // member variables.
     std::vector<Move> equal_pair;
     int equal_pair_count = 0;
 
-    int sum_uncovered_weight; // f(X), 记录未覆盖顶点的加权和, Eq (11);
+    unsigned long long sum_uncovered_weight; // f(X), 记录未覆盖顶点的加权和, Eq (11);
 
     long long iter;
